02_Cloud_App/requirements.txt
streamlit
supabase
python-dotenv
pandas
plotly
streamlit-aggrid

streamlit-pydantic
streamlit-modal
streamlit-image-select
streamlit-camera-input-live
watchdog